packages/app/src/__deprecated__/renderer/store/helpers/action.helper.ts
/**
 * Copyright (c) Mudita sp. z o.o. All rights reserved.
 * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
 */

// AUTO DISABLED - fix me if you like :)
// eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types, @typescript-eslint/no-explicit-any
export const pendingAction = (event: any): string => {
  return [event, "pending"].join("/")
}

// AUTO DISABLED - fix me if you like :)
// eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types, @typescript-eslint/no-explicit-any
export const fulfilledAction = (event: any): string => {
  return [event, "fulfilled"].join("/")
}

// AUTO DISABLED - fix me if you like :)
// eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types, @typescript-eslint/no-explicit-any
export const rejectedAction = (event: any): string => {
  return [event, "rejected"].join("/")
}
